requirements.txt
requests
google-generativeai
fan-tools
